+++ CHANGES.md
@@ -2,12 +2,13 @@
 
 ### 0.5.1
 
-* [#28] Updated: now `StreamEx.select`, `EntryStream.selectKeys` and `EntryStream.selectValues` accept any type (not necessarily the subtype of current element).
-* [#33] Updated: `StreamEx.split("", pattern)` now returns stream of single `""` string instead of empty stream.
 * [#13] Added: `StreamEx.split` to split with single character delimiter.
+* [#28] Updated: now `StreamEx.select`, `EntryStream.selectKeys` and `EntryStream.selectValues` accept any type (not necessarily the subtype of current element).
+* [#32] Added: `MoreCollectors.dominators` collector which collects the elements to the list leaving only "dominators".   
+* [#33] Updated: `StreamEx.split("", pattern)` now returns stream of single `""` string instead of empty stream.
 * [#35] Added: construction of all stream types from the `Iterator`.   
 * [#36] Added: `StreamEx.of(Enumeration)` static method.
-* [#32] Added: `MoreCollectors.dominators` collector which collects the elements to the list leaving only "dominators".   
+* [#38] Added: `scanLeft` for primitive streams.
 * Updated documentation. 
 
 ### 0.5.0