+++ CHANGES.md
@@ -1,5 +1,8 @@
 # StreamEx changes
 
+### 0.5.2
+* [#42] Added: `EntryStream.ofTree` methods to stream the tree-like structure tracking nodes depth.
+
 ### 0.5.1
 
 * [#13] Added: `StreamEx.split` to split with single character delimiter.